@@ -1,11 +1,17 @@
-const express = require('express');
-const saveUserEvent = require('./fb_functions');
-const {UserEvent, getSchema} = require('./schema');
-const { handleIncomingEvent } = require('./event_handler');
+import * as dotenv from 'dotenv';
+import express, {json} from 'express';
+import {saveUserEvent} from './fb_functions.js';
+import {handleIncomingEvent} from './user_event_handler.js';
+import {UserEvent, getSchema} from './schema.js';
+import {
+  createUserWithPassword,
+  signUserOut,
+  loginWithEmailAndPassword,
+} from './auth/email_password_auth.js';
+
+dotenv.config();
 const app = express();
-const router = express.Router();
-// Serve the files in /assets at the URI /assets.
-app.use(express.json());
+app.use(json());
 
 // Create a new UserEvent
 app.post('/events/', async (req, res) => {
@@ -13,7 +19,6 @@ app.post('/events/', async (req, res) => {
     const eventData = req.body;
     const eventType = eventData.event_type;
     console.log(req.body);
-
 
     // Get the appropriate event model based on the event type
     const EventModel = getSchema(eventType);
@@ -45,79 +50,118 @@ app.post('/events/', async (req, res) => {
   }
 });
 
+// Sign in with email and password
+app.post('/signout', async (req, res) => {
+  try {
+    const result = await signUserOut();
+    res.status(201).json({message: 'Successful user signout:', result});
+  } catch (error) {
+    console.log;
+    'Error: ', error;
+    res.status(500).json({error: 'Failed to sign user out.'});
+  }
+});
+
+// Sign user out
+app.post('/registerWithPassword', async (req, res) => {
+  try {
+    const response = req.body;
+    const email = response.email;
+    const password = response.password;
+    const uid = await createUserWithPassword(email, password);
+    res.status(201).json({message: 'Successfully created User:', uid});
+  } catch (error) {
+    console.log('Error: ', error);
+    res.status(500).json({error: 'Failed to create User'});
+  }
+});
+
+// Login in with password
+app.post('/loginWithPassword', async (req, res) => {
+  try {
+    const response = req.body;
+    const email = response.email;
+    const password = response.password;
+    const uid = await loginWithEmailAndPassword(email, password);
+    res.status(201).json({message: 'Successful Login:' + uid});
+  } catch (error) {
+    console.log('Error: ', error);
+    res.status(500).json({error: 'Failed to login'});
+  }
+});
 // Get all UserEvents of a specific type
-router.get('/events/:event_type', async (req, res) => {
-  try {
-    const eventType = req.params.event_type;
+// router.get('/events/:event_type', async (req, res) => {
+//   try {
+//     const eventType = req.params.event_type;
 
-    // Get the appropriate event model based on the event type
-    const EventModel = eventModels[eventType];
-    if (!EventModel) {
-      res.status(400).json({error: 'Invalid event type'});
-      return;
-    }
+//     // Get the appropriate event model based on the event type
+//     const EventModel = eventModels[eventType];
+//     if (!EventModel) {
+//       res.status(400).json({error: 'Invalid event type'});
+//       return;
+//     }
 
-    // Find all events of the specified type
-    const userEvents = await EventModel.find();
-    res.status(200).json(userEvents);
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({error: 'Failed to fetch UserEvents'});
-  }
-});
+//     // Find all events of the specified type
+//     const userEvents = await EventModel.find();
+//     res.status(200).json(userEvents);
+//   } catch (error) {
+//     console.error(error);
+//     res.status(500).json({error: 'Failed to fetch UserEvents'});
+//   }
+// });
 
-// Get a specific UserEvent of a specific type by ID
-app.get('/events/:event_type/:id', async (req, res) => {
-  try {
-    const eventType = req.params.event_type;
+// // Get a specific UserEvent of a specific type by ID
+// app.get('/events/:event_type/:id', async (req, res) => {
+//   try {
+//     const eventType = req.params.event_type;
 
-    // Get the appropriate event model based on the event type
-    const EventModel = eventModels[eventType];
-    if (!EventModel) {
-      res.status(400).json({error: 'Invalid event type'});
-      return;
-    }
+//     // Get the appropriate event model based on the event type
+//     const EventModel = eventModels[eventType];
+//     if (!EventModel) {
+//       res.status(400).json({error: 'Invalid event type'});
+//       return;
+//     }
 
-    const userEvent = await EventModel.findById(req.params.id);
-    if (!userEvent) {
-      res.status(404).json({error: 'UserEvent not found'});
-      return;
-    }
-    res.status(200).json(userEvent);
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({error: 'Failed to fetch UserEvent'});
-  }
-});
+//     const userEvent = await EventModel.findById(req.params.id);
+//     if (!userEvent) {
+//       res.status(404).json({error: 'UserEvent not found'});
+//       return;
+//     }
+//     res.status(200).json(userEvent);
+//   } catch (error) {
+//     console.error(error);
+//     res.status(500).json({error: 'Failed to fetch UserEvent'});
+//   }
+// });
 
-// Update a specific UserEvent of a specific type by ID
-router.put('/events/:event_type/:id', async (req, res) => {
-  try {
-    const eventType = req.params.event_type;
-    const eventData = req.body;
+// // Update a specific UserEvent of a specific type by ID
+// router.put('/events/:event_type/:id', async (req, res) => {
+//   try {
+//     const eventType = req.params.event_type;
+//     const eventData = req.body;
 
-    // Get the appropriate event model based on the event type
-    const EventModel = eventModels[eventType];
-    if (!EventModel) {
-      res.status(400).json({error: 'Invalid event type'});
-      return;
-    }
+//     // Get the appropriate event model based on the event type
+//     const EventModel = eventModels[eventType];
+//     if (!EventModel) {
+//       res.status(400).json({error: 'Invalid event type'});
+//       return;
+//     }
 
-    const userEvent = await EventModel.findByIdAndUpdate(
-      req.params.id,
-      eventData,
-      {new: true}
-    );
-    if (!userEvent) {
-      res.status(404).json({error: 'UserEvent not found'});
-      return;
-    }
-    res.status(200).json(userEvent);
-  } catch (error) {
-    console.error(error);
-    res.status(500).json({error: 'Failed to update UserEvent'});
-  }
-});
+//     const userEvent = await EventModel.findByIdAndUpdate(
+//       req.params.id,
+//       eventData,
+//       {new: true}
+//     );
+//     if (!userEvent) {
+//       res.status(404).json({error: 'UserEvent not found'});
+//       return;
+//     }
+//     res.status(200).json(userEvent);
+//   } catch (error) {
+//     console.error(error);
+//     res.status(500).json({error: 'Failed to update UserEvent'});
+//   }
+// });
 
 // Delete a specific User
 
