@@ -1,52 +1,52 @@
-const { Schema, model } = require('mongoose');
+import {Schema, model} from 'mongoose';
 
 // Define the event schemas
 const eventSchemas = {
   userRegistration: new Schema({
-    email: { type: String, required: true },
-    password: { type: String, required: true },
+    email: {type: String, required: true},
+    password: {type: String, required: true},
     // Additional fields specific to user registration
   }),
   userSignIn: new Schema({
-    status: { type: String, required: true },
-    timestamp: { type: Date, default: Date.now },
-    social: { type: Boolean, required: true },
-    social_type: { type: String, required: false },
+    status: {type: String, required: true},
+    timestamp: {type: Date, default: Date.now},
+    social: {type: Boolean, required: true},
+    social_type: {type: String, required: false},
   }),
   userWalletConnected: new Schema({
-    status: { type: String, required: true },
-    timestamp: { type: Date, default: Date.now },
-    wallet_type: { type: Boolean, required: true },
+    status: {type: String, required: true},
+    timestamp: {type: Date, default: Date.now},
+    wallet_type: {type: Boolean, required: true},
   }),
   userForgotPassword: new Schema({
-    email: { type: String, required: true },
+    email: {type: String, required: true},
   }),
   userAuthentication: new Schema({
-    email: { type: String, required: true },
-    verified: { type: Boolean, required: true },
-    timestamp: { type: Date, default: Date.now },
+    email: {type: String, required: true},
+    verified: {type: Boolean, required: true},
+    timestamp: {type: Date, default: Date.now},
   }),
   userProfileUpdate: new Schema({
-    user_id: { type: String, required: true },
-    fields: { type: Map, of: String },
+    user_id: {type: String, required: true},
+    fields: {type: Map, of: String},
   }),
   userAccountDeletion: new Schema({
-    user_id: { type: String, required: true },
-    reason: { type: String, required: false },
+    user_id: {type: String, required: true},
+    reason: {type: String, required: false},
   }),
   userInteraction: new Schema({
-    user_id: { type: String, required: true },
-    interaction: { type: String, required: true },
+    user_id: {type: String, required: true},
+    interaction: {type: String, required: true},
   }),
 };
 
 // Define the master schema
 const userEventSchema = new Schema({
-  event_id: { type: String, required: true },
-  event_type: { type: String, required: true },
-  user_id: { type: String, required: true },
-  timestamp: { type: Date, default: Date.now },
-  additional_data: { type: Schema.Types.Mixed },
+  event_id: {type: String, required: true},
+  event_type: {type: String, required: true},
+  user_id: {type: String, required: true},
+  timestamp: {type: Date, default: Date.now},
+  additional_data: {type: Schema.Types.Mixed},
 });
 
 // Define the models for the event schemas
@@ -77,8 +77,4 @@ const getSchema = (event) => {
 const UserEvent = model('UserEvent', userEventSchema);
 
 // Export the models
-module.exports = {
-  UserEvent,
-  eventSchemas,
-  getSchema
-};
+export {UserEvent, eventSchemas, getSchema};
