@@ -1,8 +1,8 @@
-import { getAuth, signInWithPopup, GoogleAuthProvider } from "firebase/auth";
+import {getAuth, signInWithPopup, GoogleAuthProvider} from 'firebase-admin';
 
 const provider = new GoogleAuthProvider();
 const auth = getAuth();
-firebase.auth().useDeviceLanguage();
+auth().useDeviceLanguage();
 
 signInWithPopup(auth, provider)
   .then((result) => {
@@ -11,15 +11,19 @@ signInWithPopup(auth, provider)
     const token = credential.accessToken;
     // The signed-in user info.
     const user = result.user;
-    // IdP data available using getAdditionalUserInfo(result)
-    // ...
-  }).catch((error) => {
+    console.log('Token: ', token);
+    console.log('User: ', user.uid);
+  })
+  .catch((error) => {
     // Handle Errors here.
-    const errorCode = error.code;
-    const errorMessage = error.message;
-    // The email of the user's account used.
-    const email = error.customData.email;
-    // The AuthCredential type that was used.
-    const credential = GoogleAuthProvider.credentialFromError(error);
+    console.error('Failed to sign in user:');
+    console.error('Error Code:', error.code);
+    console.error('Error Message:', error.message);
+    console.error('Email:', error.customData.email);
+    console.error(
+      'Credential:',
+      auth.GoogleAuthProvider.credentialFromError(error)
+    );
+    throw error;
     // ...
   });
